database-ingestion/requirements.txt
yfinance
pandas
numpy
uvicorn
fastapi
pymongo
python-dotenv
requests